LPS-63059 Move MySQL provider to its own module
