+++ modules/util/portal-tools-data-partitioning-sql-builder-mysql/.gitignore
@@ -1,0 +1,2 @@
+src/test/resources/*.properties
+third-party/*.jar
