+++ modules/util/portal-tools-data-partitioning-sql-builder-mysql/bnd.bnd
@@ -1,0 +1,4 @@
+Bundle-Name: Liferay Portal Tools Data Partitioning SQL Builder MySQL
+Bundle-SymbolicName: com.liferay.portal.tools.data.partitioning.sql.builder.mysql
+Bundle-Version: 1.0.0
+Export-Package: com.liferay.portal.tools.data.partitioning.sql.builder.db.mysql
