+++ modules/util/portal-tools-data-partitioning-sql-builder-mysql/src/main/java/com/liferay/portal/tools/data/partitioning/sql/builder/db/mysql/MySQLProvider.java
