+++ modules/util/portal-tools-data-partitioning-sql-builder-mysql/src/main/resources/sample-mysql.properties
