+++ modules/util/portal-tools-data-partitioning-sql-builder-mysql/src/test/java/com/liferay/portal/tools/data/partitioning/sql/builder/exporter/ShardExporterFactoryTest.java
@@ -14,8 +14,7 @@
 
 package com.liferay.portal.tools.data.partitioning.sql.builder.exporter;
 
-import com.liferay.portal.tools.data.partitioning.sql.builder.db.postgresql.PostgreSQLProvider;
-import com.liferay.portal.tools.data.partitioning.sql.builder.exporter.exception.DBProviderNotAvailableException;
+import com.liferay.portal.tools.data.partitioning.sql.builder.db.mysql.MySQLProvider;
 import com.liferay.portal.tools.data.partitioning.sql.builder.internal.util.PropsReader;
 
 import java.net.URL;
@@ -31,25 +30,11 @@ import org.junit.Test;
  */
 public class ShardExporterFactoryTest {
 
-	@Test(expected = DBProviderNotAvailableException.class)
-	public void testGetShardExporter() throws Exception {
-		ShardExporterFactory.getShardExporter(new Properties());
-	}
-
 	@Test
-	public void testGetShardExporterReturnsPostgreSQLProvider()
-		throws Exception {
-
-		testGetShardExporter("postgresql", PostgreSQLProvider.class);
-	}
-
-	protected void testGetShardExporter(
-			String databaseType, Class<?> providerClass)
-		throws Exception {
-
+	public void testGetShardExporterReturnsMySQLProvider() throws Exception {
 		Class<?> clazz = getClass();
 
-		URL url = clazz.getResource("/" + databaseType +".properties");
+		URL url = clazz.getResource("/mysql.properties");
 
 		Assume.assumeNotNull(url);
 
@@ -57,6 +42,8 @@ public class ShardExporterFactoryTest {
 
 		ShardExporter shardExporter = ShardExporterFactory.getShardExporter(
 			properties);
+
+		Class<MySQLProvider> providerClass = MySQLProvider.class;
 
 		Assert.assertTrue(providerClass.isInstance(shardExporter));
 	}
